ironman_airflow/requirements.txt
AIRFLOW_UID=50000
_PIP_ADDITIONAL_REQUIREMENTS=apache-airflow-providers-databricks apache-airflow-providers-amazon apache-airflow-providers-postgres pandas

